Split Srvice-class and update package
