+++ src/main/java/practice/kadai22sep7th/entity/AirportEntity.java
@@ -1,4 +1,4 @@
-package practice.kadai22sep7th;
+package practice.kadai22sep7th.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,7 +6,7 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public class AirportEntity {
-    
+
     private final String airportCode;
     private String airportName;
     private String country;
